# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Shanti Food API
  version: 1.6.0
  description: API para gestión de pedidos de arrocería vía WhatsApp

servers:
  - url: http://localhost:3000/api/v1
    description: Local development

security: []

paths:
  /auth/login:
    post:
      summary: Login del administrador (obtener JWT)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Token JWT generado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  role:
                    type: string
                    enum: [admin, delivery]
        '401':
          description: Credenciales inválidas
        '403':
          description: Dashboard de entregas deshabilitado (rol delivery con DELIVERY_DASHBOARD_ENABLED=false)

  /config/public:
    get:
      summary: Feature flags públicas (no requiere auth)
      tags: [Config]
      security: []
      responses:
        '200':
          description: Configuración pública
          content:
            application/json:
              schema:
                type: object
                properties:
                  deliveryDashboardEnabled:
                    type: boolean
                    description: "Si el dashboard de repartidores está habilitado (default: true)"

  /events:
    get:
      summary: Server-Sent Events — actualizaciones de pedidos en tiempo real
      tags: [Events]
      responses:
        '200':
          description: Stream SSE con eventos de pedidos
          content:
            text/event-stream:
              schema:
                type: string
                example: |
                  data: {"type":"connected"}

                  event: orderCreated
                  data: {"id":"SH-001","status":"pending","total":25000}

                  event: orderUpdated
                  data: {"id":"SH-001","status":"confirmed","total":25000}

  /users:
    get:
      summary: Listar usuarios (admin— filtra por rol)
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, delivery]
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          description: Token requerido

    post:
      summary: Crear usuario (admin o repartidor)
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Username ya existe

  /users/{userId}:
    patch:
      summary: Editar usuario (nombre, contraseña, estado activo)
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                password:
                  type: string
                  description: Dejar vacío para no cambiar
                active:
                  type: boolean
      responses:
        '200':
          description: Usuario actualizado
        '404':
          description: Usuario no encontrado

  /users/{userId}/stats:
    get:
      summary: Estadísticas de entregas de un repartidor
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stats del repartidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriverStats'
        '404':
          description: Usuario no encontrado

  /orders:
    post:
      summary: Crear nuevo pedido
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Pedido creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Datos inválidos
        '409':
          description: Producto no disponible

    get:
      summary: Listar pedidos (admin ve todos; repartidor ve solo los asignados a él con status=ready)
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [delivery, pickup]
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

  /orders/{orderId}/assign:
    patch:
      summary: Asignar repartidor a un pedido (solo admin)
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [driverId]
              properties:
                driverId:
                  type: integer
                  description: ID del usuario (repartidor) al que se le asigna el pedido
      responses:
        '200':
          description: Pedido actualizado con el repartidor asignado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: driverId faltante o pedido ya entregado/cancelado
        '403':
          description: No autorizado (se requiere rol admin)
        '404':
          description: Pedido no encontrado

  /orders/{orderId}/delivery-proof:
    post:
      summary: Subir foto de evidencia de entrega
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
      responses:
        '200':
          description: Foto subida
          content:
            application/json:
              schema:
                type: object
                properties:
                  delivery_proof_url:
                    type: string
        '404':
          description: Pedido no encontrado

  /orders/{orderId}:
    get:
      summary: Obtener pedido por ID
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del pedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Pedido no encontrado

    patch:
      summary: Actualizar estado del pedido
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, delivered, cancelled]
                notes:
                  type: string
      responses:
        '200':
          description: Pedido actualizado

  /products:
    get:
      summary: Listar productos del menú
      responses:
        '200':
          description: Lista de productos disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /webhooks/whatsapp:
    post:
      summary: Webhook para mensajes de WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhatsAppWebhook'
      responses:
        '200':
          description: Mensaje procesado

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtenido desde POST /auth/login

  schemas:
    OrderRequest:
      type: object
      required: [customer, items, type, paymentMethod]
      properties:
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
          minItems: 1
        type:
          type: string
          enum: [delivery, pickup]
        address:
          type: string
          description: Requerido si type es delivery
        paymentMethod:
          type: string
          enum: [cash, nequi]
        notes:
          type: string

    Order:
      type: object
      properties:
        id:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        type:
          type: string
          enum: [delivery, pickup]
        address:
          type: string
        paymentMethod:
          type: string
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        status:
          type: string
        notes:
          type: string
        deliveryProofUrl:
          type: string
          description: URL de foto de evidencia de entrega
        deliveredBy:
          type: integer
          description: ID del usuario (repartidor) que entregó el pedido
        assignedDriver:
          type: integer
          description: ID del usuario (repartidor) asignado al pedido (P7)
        createdAt:
          type: string
          format: date-time
        estimatedReadyAt:
          type: string
          format: date-time

    Customer:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
        phone:
          type: string
          pattern: '^\+?[0-9]{10,15}$'

    OrderItem:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
        customizations:
          type: array
          items:
            type: string
          description: Ej. ["sin cebolla", "picante"]
        notes:
          type: string

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
          enum: [arroz_chino, bandeja_paisa, bebidas, otros]
        price:
          type: number
        description:
          type: string
        available:
          type: boolean
        preparationMinutes:
          type: integer

    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, delivery]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time

    UserRequest:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, delivery]

    DriverStats:
      type: object
      properties:
        userId:
          type: integer
        name:
          type: string
        totalDelivered:
          type: integer
          description: Total de pedidos entregados
        totalAmount:
          type: number
          description: Suma del total de pedidos entregados
        last30Days:
          type: integer
          description: Pedidos entregados en los últimos 30 días

    WhatsAppWebhook:
      type: object
      properties:
        messageId:
          type: string
        from:
          type: string
        type:
          type: string
          enum: [text, interactive, location, order]
        text:
          type: object
          properties:
            body:
              type: string
        interactive:
          type: object
          properties:
            type:
              type: string
            buttonReply:
              type: object
            listReply:
              type: object